Sync

+++ ranked/calibration.py
@@ -1,19 +1,26 @@
 import os
 
-from orion.client import build_experiment
+try:
+    from orion.client import build_experiment
+
+    ERROR = None
+except ImportError as err:
+    ERROR = err
 
 from ranked.datasets.synthetic import SimulationConfig, create_simulated_matchups
-from ranked.models.glicko2 import Glicko2
 from ranked.models.noskill import NoSkill
 from ranked.simulation import Simulation
 
 
 def optimize(klass, max_trials=1000):
+    if ERROR is not None:
+        raise ERROR
+
     center = 1500
 
     try:
         os.remove("orion.pkl")
-    except:
+    except OSError:
         pass
 
     experiment = build_experiment(
